Updated Ubuntu installation script.
Script has two PyQt-Qwt install methods, the default being via packages and the other via source. It now detects the Linux distro as Ubuntu and exits if the distro is not Ubuntu. The script is designed to support Ubuntu versions 18.04 (package/source yet to be added), 20.04 and attempts to support other compatible versions greater than 20.04 (i.e. 21.04).

N.B. Packages and source compiles for Ubuntu 18.04 will be added soon.